feat: add code to cli and modules folders
